网站部分/mysite/requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
baidu-aip==2.2.13.0
beautifulsoup4==4.6.0
bleach==2.1.3
bosonnlp==0.11.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
charts==0.4.6
colorama==0.3.9
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
Django==2.1.7
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
dukpy==0.2.2
echarts-china-cities-pypkg==0.0.9
echarts-china-counties-pypkg==0.0.2
echarts-china-misc-pypkg==0.0.1
echarts-china-provinces-pypkg==0.0.3
echarts-cities-pypkg==0.0.1
echarts-countries-pypkg==0.1.6
entrypoints==0.2.3
future==0.17.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
javascripthon==0.10
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-echarts-pypkg==0.1.2
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
kiwisolver==1.0.1
lml==0.0.2
lxml==4.2.0
macropy3==1.1.0b2
MarkupSafe==1.0
matplotlib==3.0.2
mistune==0.8.3
mongoengine==0.17.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numpy==1.15.4
pandas==0.23.4
pandocfilters==1.4.2
parsel==1.5.1
parso==0.2.0
pickleshare==0.7.4
Pillow==5.4.0
prompt-toolkit==1.0.15
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyDispatcher==2.0.5
pyecharts==0.5.11
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
pyecharts-snapshot==0.1.10
Pygments==2.2.0
PyHamcrest==1.9.0
pymongo==3.6.1
PyMySQL==0.9.3
pyOpenSSL==18.0.0
pyparsing==2.3.1
pypiwin32==223
pyserial==3.4
python-dateutil==2.7.2
pytz==2018.7
pywin32==224
pywinpty==0.5.1
PyYAML==3.13
pyzmq==17.0.0
qtconsole==4.3.1
queuelib==1.5.0
requests==2.18.4
scipy==1.2.1
Scrapy==1.5.1
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.15.0
six==1.11.0
terminado==0.8.1
testpath==0.3.1
thulac==0.2.0
tornado==5.0.2
traitlets==4.3.2
Twisted==18.9.0
urllib3==1.22
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.1
wordcloud==1.5.0
xlrd==1.2.0
zope.interface==4.6.0
